Moved message components directory
