+++ renderer/components/messages/alias-delete.js
