+++ renderer/components/messages/avatar.js
