+++ renderer/components/messages/cert-delete.js
