+++ renderer/components/messages/deployment-delete.js
