+++ renderer/components/messages/domain-buy.js
