+++ renderer/components/messages/domain-delete.js
