+++ renderer/components/messages/domain-move-out-request-sent.js
