+++ renderer/components/messages/domain-transfer-in-canceled.js
