+++ renderer/components/messages/scale-auto.js
