+++ renderer/components/messages/secret-delete.js
