+++ renderer/components/messages/secret-rename.js
