+++ renderer/components/messages/set-scale.js
