+++ renderer/components/messages/team-avatar-update.js
