+++ renderer/components/messages/team-delete.js
