+++ renderer/components/messages/team-member-delete.js
