+++ renderer/components/messages/team-name-update.js
